+++ src/herd.rs
@@ -7,16 +7,16 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct Herd {
-    pub id: u32,
-    pub pasture_id: u32,
-    pub species_id: u32,
+    pub id: usize,
+    pub pasture_id: usize,
+    pub species_id: usize,
 
     #[serde(skip_serializing)]
     pub size: u32,
 }
 
 impl Herd {
-    pub const fn new(id: u32, pasture: &Pasture, species: &Species) -> Self {
+    pub const fn new(id: usize, pasture: &Pasture, species: &Species) -> Self {
         Herd {
             id,
             pasture_id: pasture.id,
@@ -32,7 +32,7 @@ pub fn expand_herd_vec(herds: &mut Vec<Herd>, species: &Vec<Species>, pastures: 
 
     for pasture in pastures.iter() {
         herds.push(Herd::new(
-            herds.len() as u32,
+            herds.len(),
             pasture,
             species_distribution.sample(&mut rng),
         ));
